+++ src/ui/gl/widget/HotbarWidget.cpp
@@ -10,7 +10,7 @@
 #include "ui/gl/UIContext.h"
 
 namespace Game3 {
-	HotbarWidget::HotbarWidget(double scale): scale(scale) {
+	HotbarWidget::HotbarWidget(float scale): scale(scale) {
 		for (Slot slot = 0; slot < HOTBAR_SIZE; ++slot) {
 			slotWidgets.emplace_back(std::make_shared<ItemSlotWidget>(slot, INNER_SLOT_SIZE, scale, false));
 		}
